Finished AlbumTable. Also added removed LibraryTable
